scripts/substack/requirements.txt
python-substack>=1.0.0
pyyaml>=6.0
